refactor(session): isolate the converter layer and give it a safety net
Kirk's call: encapsulation is worth a converter layer, so make the layer
explicit rather than smeared through the verbs.

types.go holds what this package returns, convert.go holds every inner-to-
outer translation, and data.go is now purely persistence. Keeping the
conversions in one file makes the cost of S2 visible instead of hidden, and
gives it a single place to be tested.

The point of the split is the test it enables. A hand-written converter does
not fail loudly -- it fails by silently dropping a field when an inner type
grows one. The projection still compiles, still returns a correctly shaped
value, and quietly omits data the host needed; nothing downstream notices,
because nothing downstream knows the field existed.

convert_test.go closes that structurally rather than field by field: for
each projected pair, every exported field on the inner type must have a
counterpart on the outer one or be listed with a reason. When the
composition adds a field, the check fails and forces the decision -- carry
it, or justify dropping it -- at the moment it is cheap. A second test
removes stale entries, since a justification for a field that no longer
exists is a comment pretending to be a decision.

One omission is recorded today: record.Entry.Audience, because naming every
viewer a beat reached would tell one player which members exist and were
present, including members never perceived.
